split tests and add basic money handler for serializer
